Add ContextualNetAddress to kaspa_utils
